Added a skeleton ShipConfirm service.
This service is the first phase of a two-phase shipment creation process. The ShipConfirm stage consists of building the shipment/packages and sending it to UPS for validation purposes. If all is well, a super-long digest string is returned, which is the parsed response of this service. The second phase involves passing the digest value back to UPS in order to actually create the shipment in the system.